Update main entry point for populating user data
This will work with the new course representation; for now (until we
get the new GQL definitions for the user-specific information).  This
is responsible for parsing and setting up the user state in the best
way for the interface to use.

Keep the old one around for now, but throw errors in case some
branches are depending on this.